requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
fastapi==0.135.1
frozenlist==1.8.0
fsspec==2026.2.0
gotrue==1.3.1
greenlet==3.3.2
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==0.17.3
httpx==0.24.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.12
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.18
markdown-it-py==4.0.0
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
openai==2.28.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
postgrest==0.11.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
PyPDF2==3.0.1
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
realtime==1.0.6
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
storage3==0.6.1
StrEnum==0.4.15
strictyaml==1.7.3
supabase==1.2.0
supabase-auth==2.28.2
supabase-functions==2.28.2
supafunc==0.2.3
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
websockets==12.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
